use lockfile for java dependency inference parser
[ci skip-rust]
