+++ 3rdparty/jvm/com/github/javaparser/BUILD
@@ -1,10 +1,11 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
+# Note: This target must be kept in sync with `JavaParserTool`.
 jvm_artifact(
     name="javaparser-symbol-solver-core",
     group="com.github.javaparser",
     artifact="javaparser-symbol-solver-core",
     version="3.23.0",
-    resolve="java_parser",
+    resolve="java_parser_dev",
 )
