+++ src/python/pants/backend/java/dependency_inference/java_parser.py
@@ -8,22 +8,50 @@ import logging
 import os.path
 from dataclasses import dataclass
 
-from pants.backend.java.dependency_inference import java_parser_launcher
-from pants.backend.java.dependency_inference.java_parser_launcher import (
-    JavaParserCompiledClassfiles,
-    java_parser_artifact_requirements,
-)
+import pkg_resources
+
 from pants.backend.java.dependency_inference.types import JavaSourceDependencyAnalysis
+from pants.core.goals.generate_lockfiles import GenerateToolLockfileSentinel
 from pants.core.util_rules.source_files import SourceFiles
-from pants.engine.fs import AddPrefix, Digest, DigestContents
-from pants.engine.process import FallibleProcessResult, ProcessExecutionFailure
+from pants.engine.fs import AddPrefix, CreateDigest, Digest, DigestContents, Directory, FileContent
+from pants.engine.internals.native_engine import MergeDigests, RemovePrefix
+from pants.engine.process import FallibleProcessResult, ProcessExecutionFailure, ProcessResult
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
+from pants.engine.unions import UnionRule
 from pants.jvm.jdk_rules import InternalJdk, JvmProcess
 from pants.jvm.resolve.coursier_fetch import ToolClasspath, ToolClasspathRequest
+from pants.jvm.resolve.jvm_tool import GenerateJvmLockfileFromTool, JvmToolBase
 from pants.option.global_options import ProcessCleanupOption
+from pants.util.docutil import git_url
 from pants.util.logging import LogLevel
 
 logger = logging.getLogger(__name__)
+
+
+_LAUNCHER_BASENAME = "PantsJavaParserLauncher.java"
+
+
+class JavaParserTool(JvmToolBase):
+    options_scope = "java-parser"
+    name = "Java Parser for Dependency Inference"
+    help = "Java Parser for Dependency Inference"
+
+    default_version = "<n/a>"
+    default_artifacts = (
+        "com.fasterxml.jackson.core:jackson-databind:2.12.4",
+        "com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.12.4",
+        "com.github.javaparser:javaparser-symbol-solver-core:3.23.0",
+    )
+    default_lockfile_resource = (
+        "pants.backend.java.dependency_inference",
+        "java_parser.lock",
+    )
+    default_lockfile_path = "src/python/pants/backend/java/dependency_inference/java_parser.lock"
+    default_lockfile_url = git_url(default_lockfile_path)
+
+
+class JavaParserToolLockfileSentinel(GenerateToolLockfileSentinel):
+    resolve_name = JavaParserTool.options_scope
 
 
 @dataclass(frozen=True)
@@ -34,6 +62,11 @@ class JavaSourceDependencyAnalysisRequest:
 @dataclass(frozen=True)
 class FallibleJavaSourceDependencyAnalysisResult:
     process_result: FallibleProcessResult
+
+
+@dataclass(frozen=True)
+class JavaParserCompiledClassfiles:
+    digest: Digest
 
 
 @rule(level=LogLevel.DEBUG)
@@ -85,10 +118,13 @@ async def analyze_java_source_dependencies(
     processorcp_relpath = "__processorcp"
     toolcp_relpath = "__toolcp"
 
-    (tool_classpath, prefixed_source_files_digest,) = await MultiGet(
+    parser_lockfile_request = await Get(
+        GenerateJvmLockfileFromTool, JavaParserToolLockfileSentinel()
+    )
+    tool_classpath, prefixed_source_files_digest = await MultiGet(
         Get(
             ToolClasspath,
-            ToolClasspathRequest(artifact_requirements=java_parser_artifact_requirements()),
+            ToolClasspathRequest(lockfile=parser_lockfile_request),
         ),
         Get(Digest, AddPrefix(source_files.snapshot.digest, source_prefix)),
     )
@@ -125,8 +161,82 @@ async def analyze_java_source_dependencies(
     return FallibleJavaSourceDependencyAnalysisResult(process_result=process_result)
 
 
+def _load_javaparser_launcher_source() -> bytes:
+    return pkg_resources.resource_string(__name__, _LAUNCHER_BASENAME)
+
+
+# TODO(13879): Consolidate compilation of wrapper binaries to common rules.
+@rule
+async def build_processors(jdk: InternalJdk) -> JavaParserCompiledClassfiles:
+    dest_dir = "classfiles"
+    parser_lockfile_request = await Get(
+        GenerateJvmLockfileFromTool, JavaParserToolLockfileSentinel()
+    )
+    materialized_classpath, source_digest = await MultiGet(
+        Get(
+            ToolClasspath,
+            ToolClasspathRequest(prefix="__toolcp", lockfile=parser_lockfile_request),
+        ),
+        Get(
+            Digest,
+            CreateDigest(
+                [
+                    FileContent(
+                        path=_LAUNCHER_BASENAME,
+                        content=_load_javaparser_launcher_source(),
+                    ),
+                    Directory(dest_dir),
+                ]
+            ),
+        ),
+    )
+
+    merged_digest = await Get(
+        Digest,
+        MergeDigests(
+            (
+                materialized_classpath.digest,
+                source_digest,
+            )
+        ),
+    )
+
+    process_result = await Get(
+        ProcessResult,
+        JvmProcess(
+            jdk=jdk,
+            classpath_entries=[f"{jdk.java_home}/lib/tools.jar"],
+            argv=[
+                "com.sun.tools.javac.Main",
+                "-cp",
+                ":".join(materialized_classpath.classpath_entries()),
+                "-d",
+                dest_dir,
+                _LAUNCHER_BASENAME,
+            ],
+            input_digest=merged_digest,
+            output_directories=(dest_dir,),
+            description=f"Compile {_LAUNCHER_BASENAME} import processors with javac",
+            level=LogLevel.DEBUG,
+            # NB: We do not use nailgun for this process, since it is launched exactly once.
+            use_nailgun=False,
+        ),
+    )
+    stripped_classfiles_digest = await Get(
+        Digest, RemovePrefix(process_result.output_digest, dest_dir)
+    )
+    return JavaParserCompiledClassfiles(digest=stripped_classfiles_digest)
+
+
+@rule
+def generate_java_parser_lockfile_request(
+    _: JavaParserToolLockfileSentinel, tool: JavaParserTool
+) -> GenerateJvmLockfileFromTool:
+    return GenerateJvmLockfileFromTool.create(tool)
+
+
 def rules():
     return [
         *collect_rules(),
-        *java_parser_launcher.rules(),
+        UnionRule(GenerateToolLockfileSentinel, JavaParserToolLockfileSentinel),
     ]
